+++ lib/product_matching_v2/domain/product_match_evidence.dart
@@ -7,6 +7,8 @@ class ProductMatchEvidence {
     required Iterable<String> matchedTokens,
     required this.exactNormalizedMatch,
     required this.discoverySource,
+    this.matchedCatalogText,
+    this.matchedAlias,
   }) : matchedTokens = List.unmodifiable(matchedTokens);
 
   factory ProductMatchEvidence.fromJson(Map<String, Object?> json) =>
@@ -18,6 +20,8 @@ class ProductMatchEvidence {
         discoverySource: ProductMatchDiscoverySource.values.byName(
           json['discoverySource']! as String,
         ),
+        matchedCatalogText: json['matchedCatalogText'] as String?,
+        matchedAlias: json['matchedAlias'] as String?,
       );
 
   final String normalizedQuery;
@@ -25,12 +29,16 @@ class ProductMatchEvidence {
   final List<String> matchedTokens;
   final bool exactNormalizedMatch;
   final ProductMatchDiscoverySource discoverySource;
+  final String? matchedCatalogText;
+  final String? matchedAlias;
 
-  Map<String, Object> toJson() => {
+  Map<String, Object?> toJson() => {
         'normalizedQuery': normalizedQuery,
         'normalizedCatalogText': normalizedCatalogText,
         'matchedTokens': matchedTokens,
         'exactNormalizedMatch': exactNormalizedMatch,
         'discoverySource': discoverySource.name,
+        'matchedCatalogText': matchedCatalogText,
+        'matchedAlias': matchedAlias,
       };
 }